requirements.txt
fastapi
uvicorn
pymongo
pymongo[srv]
passlib
passlib[bcrypt]
python-jose[cryptography]